Adição de validações em português, correções no atributo next_url da paginação e mudança do atributo preco para double

+++ database/migrations/2017_03_12_195732_create_veiculos_table.php
@@ -18,7 +18,7 @@ class CreateVeiculosTable extends Migration
             $table->string('marca');
             $table->string('modelo');
             $table->integer('ano');
-            $table->float('preco');
+            $table->double('preco');
             $table->timestamps();
         });
     }
